encryption_api/requirements.txt
amqp
celery
Django==2.2.15
django-celery-results
django-jinja==2.6.0
django-logging-json==1.15
flower==0.9.3
httplib2==0.9.2
humanize==2.6.0
Jinja2==2.11.1
jsonschema==3.2.0
paramiko==2.7.2
pathlib2==2.3.5
psycopg2
PyJWT
pytest-shutil==1.7.0
PyYAML==5.3
requests==2.23.0
rpyc==4.1.4
scp==0.13.2
SQLAlchemy==1.3.19
sqlparse==0.3.1
urllib3==1.25.10
uuid==1.30
python-irodsclient @ git+https://github.com/lexis-project/python-irodsclient.git@openid_20201105
